Add role-based access control

+++ inverno-ticket/src/main/java/io/inverno/app/ticket/internal/rest/v1/PlanWebController.java
@@ -16,17 +16,20 @@
 package io.inverno.app.ticket.internal.rest.v1;
 
 import io.inverno.core.annotation.Bean;
-import io.inverno.app.ticket.internal.exception.PlanAlreadyExistsException;
 import io.inverno.app.ticket.internal.model.Plan;
 import io.inverno.app.ticket.internal.model.Ticket;
 import io.inverno.app.ticket.internal.rest.DtoMapper;
 import io.inverno.app.ticket.internal.rest.v1.dto.PlanDto;
 import io.inverno.app.ticket.internal.service.PlanService;
 import io.inverno.mod.base.resource.MediaTypes;
+import io.inverno.mod.http.base.ForbiddenException;
 import io.inverno.mod.http.base.Method;
 import io.inverno.mod.http.base.NotFoundException;
 import io.inverno.mod.http.base.Status;
 import io.inverno.mod.http.base.header.Headers;
+import io.inverno.mod.security.accesscontrol.RoleBasedAccessController;
+import io.inverno.mod.security.http.context.SecurityContext;
+import io.inverno.mod.security.identity.Identity;
 import io.inverno.mod.web.base.annotation.Body;
 import io.inverno.mod.web.base.annotation.FormParam;
 import io.inverno.mod.web.base.annotation.PathParam;
@@ -67,17 +70,25 @@ public class PlanWebController {
 	 * @return {@inverno.web.status 201} the created plan
 	 */
 	@WebRoute( method = Method.POST, consumes = MediaTypes.APPLICATION_JSON, produces = MediaTypes.APPLICATION_JSON )
-	public Mono<PlanDto> createPlan(@Body PlanDto plan, WebExchange<?> exchange) {
-		plan.setId(null);
-		return this.planDtoMapper.toDomain(plan)
-			.flatMap(this.planService::savePlan)
-			.doOnNext(savedPlan -> exchange.response()
-				.headers(headers -> headers
-					.status(Status.CREATED)
-					.add(Headers.NAME_LOCATION, exchange.request().getPathBuilder().segment(savedPlan.getId().toString()).buildPath())
-				)
-			)
-			.flatMap(this.planDtoMapper::toDto);
+	public Mono<PlanDto> createPlan(@Body PlanDto plan, WebExchange<? extends SecurityContext<? extends Identity, ? extends RoleBasedAccessController>> exchange) {
+		return exchange.context().getAccessController()
+			.orElseThrow(() -> new ForbiddenException("Missing access controller"))
+			.hasRole("admin")
+			.flatMap(hasRole -> {
+				if(!hasRole) {
+					throw new ForbiddenException();
+				}
+				plan.setId(null);
+				return this.planDtoMapper.toDomain(plan)
+					.flatMap(this.planService::savePlan)
+					.doOnNext(savedPlan ->
+						exchange.response().headers(headers -> headers
+							.status(Status.CREATED)
+							.add(Headers.NAME_LOCATION, exchange.request().getPathBuilder().segment(savedPlan.getId().toString()).buildPath())
+						)
+					)
+					.flatMap(this.planDtoMapper::toDto);
+			});
 	}
 
 	/**
@@ -105,7 +116,7 @@ public class PlanWebController {
 	public Mono<PlanDto> getPlan(@PathParam long planId, @QueryParam Optional<List<Ticket.Status>> statuses) {
 		return statuses.map(s -> this.planService.getPlan(planId, s)).orElse(this.planService.getPlan(planId))
 			.flatMap(this.planDtoMapper::toDto)
-			.switchIfEmpty(Mono.error(() -> new NotFoundException()));
+			.switchIfEmpty(Mono.error(NotFoundException::new));
 	}
 
 	/**
@@ -117,12 +128,20 @@ public class PlanWebController {
 	 * @return the updated plan
 	 */
 	@WebRoute( path = "/{planId}", method = Method.PUT, consumes = MediaTypes.APPLICATION_JSON, produces = MediaTypes.APPLICATION_JSON )
-	public Mono<PlanDto> updatePlan(@PathParam long planId, @Body PlanDto plan) {
-		plan.setId(planId);
-		return this.planDtoMapper.toDomain(plan)
-			.flatMap(this.planService::savePlan)
-			.flatMap(this.planDtoMapper::toDto)
-			.switchIfEmpty(Mono.error(() -> new NotFoundException()));
+	public Mono<PlanDto> updatePlan(@PathParam long planId, @Body PlanDto plan, SecurityContext<? extends Identity, ? extends RoleBasedAccessController> securityContext) {
+		return securityContext.getAccessController()
+			.orElseThrow(() -> new ForbiddenException("Missing access controller"))
+			.hasRole("admin")
+			.flatMap(hasRole -> {
+				if (!hasRole) {
+					throw new ForbiddenException();
+				}
+				plan.setId(planId);
+				return this.planDtoMapper.toDomain(plan)
+					.flatMap(this.planService::savePlan)
+					.flatMap(this.planDtoMapper::toDto)
+					.switchIfEmpty(Mono.error(NotFoundException::new));
+			});
 	}
 
 	/**
@@ -134,10 +153,18 @@ public class PlanWebController {
 	 * @throws NotFoundException if there's no ticket with the specified id
 	 */
 	@WebRoute( path = "/{planId}", method = Method.DELETE, produces = MediaTypes.APPLICATION_JSON )
-	public Mono<PlanDto> deletePlan(@PathParam long planId) {
-		return this.planService.removePlan(planId)
-			.flatMap(this.planDtoMapper::toDto)
-			.switchIfEmpty(Mono.error(() -> new NotFoundException()));
+	public Mono<PlanDto> deletePlan(@PathParam long planId, SecurityContext<? extends Identity, ? extends RoleBasedAccessController> securityContext) {
+		return securityContext.getAccessController()
+			.orElseThrow(() -> new ForbiddenException("Missing access controller"))
+			.hasRole("admin")
+			.flatMap(hasRole -> {
+				if (!hasRole) {
+					throw new ForbiddenException();
+				}
+				return this.planService.removePlan(planId)
+					.flatMap(this.planDtoMapper::toDto)
+					.switchIfEmpty(Mono.error(NotFoundException::new));
+			});
 	}
 
 	/**
